refactor: sponsor feature
